tests/wpt/tests/workers/Worker_NosniffErr.htm

<!DOCTYPE html>
<title> Worker with nosniff X-Content-Type-Options header </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function(t) {
  var worker = new Worker("./support/nosiniff-error-worker.py");
  worker.onerror = t.step_func_done(function(e) {
    assert_equals(e.type, 'error');
  });
});
</script>
