LayoutTests/imported/w3c/web-platform-tests/workers/Worker_NosniffErr-expected.txt

CONSOLE MESSAGE: Worker failed to load. Refused to execute http://localhost:8800/workers/support/nosiniff-error-worker.py as script because "X-Content-Type-Options: nosniff" was given and its Content-Type is not a script MIME type.

PASS  Worker with nosniff X-Content-Type-Options header
